Refactor: city 마커 표시 함수 임시 분리
